% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arrow.R
\docType{class}
\name{Arrow}
\alias{Arrow}
\title{Arrow}
\description{
An R6 class to represent an directed edge in a digraph.
}
\details{
Arrows are the formal term for links between pairs of nodes in a
graph. It is not intended that package users creating models should 
instantiate the `Arrow` class. Instead, it is included in the
package as a convenience class used in the construction and traversal
of models by the package methods themselves, and for examining
and printing model structure.
}
\author{
Andrew J. Sims \email{andrew.sims@newcastle.ac.uk}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Arrow$new()}}
\item \href{#method-get_source}{\code{Arrow$get_source()}}
\item \href{#method-get_target}{\code{Arrow$get_target()}}
\item \href{#method-get_label}{\code{Arrow$get_label()}}
\item \href{#method-clone}{\code{Arrow$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an object of type 'Arrow'.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Arrow$new(source, target, label = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{Node from which the arrow emerges.}

\item{\code{target}}{second Node to which the arrow points.}

\item{\code{label}}{Character string containing the arrow label.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Arrow` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_source"></a>}}
\if{latex}{\out{\hypertarget{method-get_source}{}}}
\subsection{Method \code{get_source()}}{
Access source node.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Arrow$get_source()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
`Node` from which the arrow leads.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_target"></a>}}
\if{latex}{\out{\hypertarget{method-get_target}{}}}
\subsection{Method \code{get_target()}}{
Access target node.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Arrow$get_target()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
`Node` to which the arrow points.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_label"></a>}}
\if{latex}{\out{\hypertarget{method-get_label}{}}}
\subsection{Method \code{get_label()}}{
Access label.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Arrow$get_label()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Label of the arrow; character string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Arrow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
