% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstModVar.R
\docType{class}
\name{ConstModVar}
\alias{ConstModVar}
\title{ConstModVar}
\description{
An R6 class for a constant in a model
}
\details{
A ModVar with no uncertainty in its value. It has no
distribution and there are no hyperparameters. Its 
benefit over using a regular 'numeric' variable in a model is that
it will appear in automatic tabulations of the model variables
associated with a model and therefore be explicitly documented
as a model input.
}
\author{
Andrew Sims \email{andrew.sims@newcastle.ac.uk}
}
\section{Super class}{
\code{\link[rdecision:ModVar]{rdecision::ModVar}} -> \code{ConstModVar}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConstModVar$new()}}
\item \href{#method-sample}{\code{ConstModVar$sample()}}
\item \href{#method-getDistribution}{\code{ConstModVar$getDistribution()}}
\item \href{#method-getMean}{\code{ConstModVar$getMean()}}
\item \href{#method-getSD}{\code{ConstModVar$getSD()}}
\item \href{#method-getQuantile}{\code{ConstModVar$getQuantile()}}
\item \href{#method-clone}{\code{ConstModVar$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModVar" data-id="get_environment">}\href{../../rdecision/html/ModVar.html#method-get_environment}{\code{rdecision::ModVar$get_environment()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModVar" data-id="get_label">}\href{../../rdecision/html/ModVar.html#method-get_label}{\code{rdecision::ModVar$get_label()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModVar" data-id="getDescription">}\href{../../rdecision/html/ModVar.html#method-getDescription}{\code{rdecision::ModVar$getDescription()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModVar" data-id="getOperands">}\href{../../rdecision/html/ModVar.html#method-getOperands}{\code{rdecision::ModVar$getOperands()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModVar" data-id="getUnits">}\href{../../rdecision/html/ModVar.html#method-getUnits}{\code{rdecision::ModVar$getUnits()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModVar" data-id="isExpression">}\href{../../rdecision/html/ModVar.html#method-isExpression}{\code{rdecision::ModVar$isExpression()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModVar" data-id="r">}\href{../../rdecision/html/ModVar.html#method-r}{\code{rdecision::ModVar$r()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModVar" data-id="set_label">}\href{../../rdecision/html/ModVar.html#method-set_label}{\code{rdecision::ModVar$set_label()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModVar" data-id="tabulate">}\href{../../rdecision/html/ModVar.html#method-tabulate}{\code{rdecision::ModVar$tabulate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModVar" data-id="unset_label">}\href{../../rdecision/html/ModVar.html#method-unset_label}{\code{rdecision::ModVar$unset_label()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModVar" data-id="value">}\href{../../rdecision/html/ModVar.html#method-value}{\code{rdecision::ModVar$value()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new constant model variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstModVar$new(description, units, const)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{A character string description of the variable
and its role in the
model. This description will be used in a tabulation of the
variables linked to a model.}

\item{\code{units}}{A character string description of the units, e.g. 'GBP',
'per year'.}

\item{\code{const}}{The constant numerical value of the object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new ConstModVar object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
Not applicable for a constant variable, no action is taken on sampling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstModVar$sample(expected = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expected}}{Logical; ignored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated ConstModVar object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDistribution"></a>}}
\if{latex}{\out{\hypertarget{method-getDistribution}{}}}
\subsection{Method \code{getDistribution()}}{
Accessor function for the name of the uncertainty distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstModVar$getDistribution()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Distribution name as character string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMean"></a>}}
\if{latex}{\out{\hypertarget{method-getMean}{}}}
\subsection{Method \code{getMean()}}{
Return the expected value of the distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstModVar$getMean()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Expected value as a numeric value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSD"></a>}}
\if{latex}{\out{\hypertarget{method-getSD}{}}}
\subsection{Method \code{getSD()}}{
Return the standard deviation of the distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstModVar$getSD()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Standard deviation as a numeric value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getQuantile"></a>}}
\if{latex}{\out{\hypertarget{method-getQuantile}{}}}
\subsection{Method \code{getQuantile()}}{
Quantiles of the uncertainty distribution; for a constant all
quantiles are returned as the value of the constant.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstModVar$getQuantile(probs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{probs}}{Numeric vector of probabilities, each in range [0,1].}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of numeric values of the same length as `probs`.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConstModVar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
