% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{dec.hurwicz}
\alias{dec.hurwicz}
\title{Hurwicz criteria}
\usage{
dec.hurwicz(m, a = 0.5)
}
\arguments{
\item{m}{matrix with strategies in rows and states in columns. Benefits in 
positive and costs in negative}

\item{a}{index of optimistic (between 0 and 1). Default: 0.5}
}
\value{
A list with best Hurwicz criteria decision and calculation matrix
}
\description{
Hurwicz criteria
}
\examples{
#
m <- data.frame(c(200, 260, 272), 
                c(300, 320, 280), 
                c(340, 320, 300))
dimnames(m) <- list(LETTERS[1:nrow(m)], 
                    paste0('State_', as.roman(1:ncol(m))))
dec.hurwicz(m, a=0.6)
}
\author{
Ruben Escribano-Garcia \email{escribano.engineer@gmail.com}
}
