% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\docType{class}
\name{DecisionTree}
\alias{DecisionTree}
\title{DecisionTree}
\value{
Updated DecisionNode object.
}
\description{
An R6 class to represent a decision tree
}
\details{
A class to represent a decision tree. An object contains a tree of
decision nodes, chance nodes and states, connected by edges, and
meeting the criteria for a decision tree.
}
\author{
Andrew J. Sims \email{andrew.sims@newcastle.ac.uk}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DecisionTree$new()}}
\item \href{#method-update}{\code{DecisionTree$update()}}
\item \href{#method-evaluatePathways}{\code{DecisionTree$evaluatePathways()}}
\item \href{#method-evaluateChoices}{\code{DecisionTree$evaluateChoices()}}
\item \href{#method-clone}{\code{DecisionTree$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new decision tree. The tree must start with a decision node
which is the root of a tree of chance nodes and state nodes
connected by edges.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DecisionTree$new(dn)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dn}}{An object inheriting from class DecisionNode.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A DecisionTree object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Update the tree by sampling model variables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DecisionTree$update(expected = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expected}}{if TRUE set model variables to their expectation; 
otherwise sample from their uncertainty distributions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-evaluatePathways"></a>}}
\if{latex}{\out{\hypertarget{method-evaluatePathways}{}}}
\subsection{Method \code{evaluatePathways()}}{
Evaluate a decision. Starting with this decision node, the function
works though all possible paths and computes the probability,
cost and utility of each.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DecisionTree$evaluatePathways(expected = TRUE, uncorrelate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expected}}{If TRUE, evaluate each model variable as its mean value,
otherwise sample each one from their uncertainty distrbution.}

\item{\code{uncorrelate}}{If TRUE, resample and update the tree between
the evaluation of each choice. This causes any model variables that
are common to more than one choice to be resampled between choices,
and removes correlation due to shared model variables. Other forms
of correlation may not be removed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame with one row per path and columns organized as
follows:
\describe{
\item{Choice}{The choice with which the path is associated.}
\item{Pathway}{The leaf node on which the pathway ends; normally the 
clinical outcome.}
\item{Probability}{The probability of traversing the pathway. The total
probability of each choice should sum to unity; i.e. the sum of the 
Probability column should equal the number of branches leaving the 
decision node.}
\item{Cost}{The cost of traversing the pathway.}
\item{ExpectedCost}{Cost \eqn{*} probability of traversing the pathway.}
\item{Utility}{The utility associated with the outcome.}
\item{ExpectedUtility}{Utility \eqn{*} probability of traversing the pathway.}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-evaluateChoices"></a>}}
\if{latex}{\out{\hypertarget{method-evaluateChoices}{}}}
\subsection{Method \code{evaluateChoices()}}{
Evaluate each choice. Starting with this decision node, the function
works though all possible paths and computes the probability,
cost and utility of each, then aggregates by choice.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DecisionTree$evaluateChoices(expected = TRUE, uncorrelate = FALSE, N = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expected}}{If TRUE, evaluate each model variable as its mean value,
otherwise sample each one from their uncertainty distrbution.}

\item{\code{uncorrelate}}{If TRUE, resample and update the tree between
the evaluation of each choice. This causes any model variables that
are common to more than one choice to be resampled between choices,
and removes correlation due to shared model variables.}

\item{\code{N}}{Number of replicates. Intended for use with PSA (expected=F);
use with expected=T will be repetitive and uninformative.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame with one row per choice per run and columns
organized as follows:
\describe{
\item{Run}{The run number}
\item{Choice}{The choice.}
\item{Cost}{Aggregate cost of the choice.}
\item{Utility}{Aggregate utility of the choice.}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DecisionTree$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
