% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{decs.summary}
\alias{decs.summary}
\title{Summary of all criterias}
\usage{
decs.summary(m, a = 0.5)
}
\arguments{
\item{m}{matrix with strategies in rows and states in columns. Benefits in 
positive and costs in negative}

\item{a}{index of optimistic for Hurwicz criteria (between 0 and 1). Default: 0.5}
}
\value{
A list matrix with best decisions for all criterias
}
\description{
Summary of all criterias
}
\examples{
#
m <- data.frame(c(200, 260, 272), 
                c(300, 320, 280), 
                c(340, 320, 300))
dimnames(m) <- list(LETTERS[1:nrow(m)], 
                    paste0('State_', as.roman(1:ncol(m))))
decs.summary(m, a=0.6)
}
\author{
Ruben Escribano-Garcia \email{escribano.engineer@gmail.com}
}
