% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des.R
\name{des}
\alias{des}
\title{Discrete event simulation solver of Markov models}
\usage{
des(
  nStates,
  nGroups = 0,
  nPatients,
  nCyclesPerYear,
  nCycles,
  state,
  group = NA,
  prevalence,
  Ip,
  Tp,
  Gp = NA,
  discount = 0,
  stub = NA,
  continue = F
)
}
\arguments{
\item{nStates}{Number of states in Markov model}

\item{nGroups}{Number of groups (default zero); states can be grouped
together and a maximum cycle limit set on group occupancy.}

\item{nPatients}{Number of patients to simulate}

\item{nCyclesPerYear}{Number of cycles per year}

\item{nCycles}{Number of time cycles to simulate}

\item{state}{Data frame of length nStates with named fields (see: State)}

\item{group}{Data frame of length nGroups with named fields (see: Group)}

\item{prevalence}{Array of length nStates which contains values for the
initial state occupancy (prevalence). All values should
be in the interval [0,1] and the sum of the elements
should be 1.}

\item{Ip}{Matrix of ANNUAL rates of transition between states. The
dimensions are (nStates,nStates) with each entry being the
annual rate of transitions from the row state to the
column state. Values are in the range [0,1]. Transitions
to self (i.e. leading diagonal) should NOT be included.}

\item{Tp}{Matrix of dimension (nStates,nStates) which contain time
independent probabilities of transitions from time limited
states to other states after time expiry. Values are in
the interval [0,1].}

\item{Gp}{Matrix of dimension (nGroups, nStates) which contain time
independent probabilities of transitions from time limited
groups to other states after time expiry. Values are in
the interval [0,1]. It is an error if any of the transitions
from a group to a state within that group are > 0.}

\item{discount}{Annual discount rate for costs and benefits as a
percentage figure, eg 3.5. This is applied to each time
cycle with log correction to convert from annual rate to
per-cycle rate if required. Default 0.00.}

\item{stub}{File stub to be used for output files, eg 'mymarkov'
would cause files called 'mymarkov-log.txt' etc to be
created. Default NA means no files are written.}

\item{continue}{If no, states are populated according to the given
prevalence; if yes, the simulation restarts in its
previous state by reading in the csv files, and the
prevalences are ignored. If stub is NA (file output is
suppressed) continue=T is disallowed.}
}
\value{
A list of 4 matrices, for state populations, costs, entries
       and utilities.
}
\description{
\code{des} is solver function for the \code{rdecision}
package for discrete event simulations. The function is based on the model
described by Sonnenberg
and Beck. This function solves a Markov state model using a Monte-Carlo
method, based on random number generation.
}
\note{
The following files will be created by the function:
'stub'-log.txt: a text file listing inputs, progress and final results
                of the simulation.
'stub'-pop.csv: a comma separated value file containing the population
                of each state at the end of each cycle.
'stub'-psp.csv: a comma separated value file containing the cumulative
                cost associated with entry and occupancy of each state
                after each cycle.
'stub'-ent.csv: a comma separated value file containing the cumulative
                number of entries into each state, after each cycle.
'stub'-utl.csv: a comma separated value file containing the cumulative
                utlity of each state, after each cycle.
}
\section{State}{

Data frame \code{state} should have the following fields:
\code{name} string which describes state (used in log output)
\code{hasCycleLimit} value TRUE if cycle limit applies (ie temporary/tunnel
                     state); FALSE otherwise (ie absorbing or normal state)
\code{cycleLimit} (integer) maximum number of cycles for which a
                  single patient can occupy the state, if
                  hasCycleLimit = TRUE. For temporary
                  states use hasCycleLimit=TRUE and cycleLimit=0.
\code{annualCost} cost of being in the state for 1 year
\code{entryCost} one-off cost associated with entering the state
\code{utility} (incremental) utility of being in the state for one year
\code{group} group number to which this state belongs (0 if no group,
             >= 1 otherwise)
}

\section{Group}{

Date frame \code{group} should have the following fields:
\code{number} (integer) group number as an integer (referenced
               by state$group); NA if not defined.
\code{name} string which describes the group name (in output)
\code{hasCycleLimit} value TRUE if cycle limit applies to the
                       group, FALSE otherwise.
\code{cycleLimit} maximum number of cycles for which a single
                  patient can occupy the group.
}

\references{
Sonnenberg FA and Beck JR, "Markov models in medical decision
making: a practical guide", Medical Decision Making 1993;13:322-339.
}
