% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fprintf.R
\name{fprintf}
\alias{fprintf}
\title{A version of fprintf for R}
\usage{
fprintf(con, fmt, ...)
}
\arguments{
\item{con}{an R connection, opened with the \link{open} command
or one its overloaded methods, such as \link{file} or \link{url}.}

\item{fmt}{a format string, of the syntax accepted by the R function
\link{sprintf}.}

\item{...}{a list of arguments, matching those specified in the
\code{fmt} argument,, to be written to the connection.}
}
\value{
The number of characters written to the connection.
}
\description{
\code{fprintf} writes a formatted string to the file open on
connection \code{con}. It is intended to be equivalent to
the Matlab and C function of the same name.
}
\examples{
 
ofid <- file("marvin.txt", open='wt')
fprintf(ofid, "The answer is \%i\n", 42)
close(ofid)
}
\seealso{
\link{sprintf}
}
