% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaModelVariable.R
\docType{class}
\name{GammaModelVariable}
\alias{GammaModelVariable}
\title{GammaModelVariable}
\description{
An R6 class for a model variable with Gamma function uncertainty
}
\details{
A model variable for which the uncertainty in the point estimate can
be modelled with a Gamma distribution. The hyperparameters of the
distribution are the shape (`alpha`) and the scale (`beta`) of
the uncertainty distribution. Note that this variable naming 
convention follos Briggs et al; `beta` is more usually used to
describe the rate parameter (reciprocal of scale.)
}
\author{
Andrew J. Sims \email{andrew.sims5@nhs.net}
}
\section{Super class}{
\code{\link[rdecision:ModelVariable]{rdecision::ModelVariable}} -> \code{GammaModelVariable}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GammaModelVariable$new()}}
\item \href{#method-sample}{\code{GammaModelVariable$sample()}}
\item \href{#method-getDistribution}{\code{GammaModelVariable$getDistribution()}}
\item \href{#method-getMean}{\code{GammaModelVariable$getMean()}}
\item \href{#method-getSD}{\code{GammaModelVariable$getSD()}}
\item \href{#method-getQuantile}{\code{GammaModelVariable$getQuantile()}}
\item \href{#method-clone}{\code{GammaModelVariable$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModelVariable" data-id="getDescription">}\href{../../rdecision/html/ModelVariable.html#method-getDescription}{\code{rdecision::ModelVariable$getDescription()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModelVariable" data-id="getUnits">}\href{../../rdecision/html/ModelVariable.html#method-getUnits}{\code{rdecision::ModelVariable$getUnits()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModelVariable" data-id="value">}\href{../../rdecision/html/ModelVariable.html#method-value}{\code{rdecision::ModelVariable$value()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an object of class GammaModelVariable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GammaModelVariable$new(description, units, alpha, beta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{A character string describing the variable.}

\item{\code{units}}{Units of the variable, as character string.}

\item{\code{alpha}}{shape parameter of the Gamma distribution.}

\item{\code{beta}}{scale parameter of the Gamma distribution.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class GammaModelVariable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
Set the value of the model variable from its uncertainty distribution.
Nothing is returned; the sampled value is returned at the next
call to `value()`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GammaModelVariable$sample(expected = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expected}}{Logical; if TRUE sets the value of the model variable
returned at subsequent calls to `value()` to be equal to the 
expectation of the variable. Default is FALSE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated GammaModelVariable object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDistribution"></a>}}
\if{latex}{\out{\hypertarget{method-getDistribution}{}}}
\subsection{Method \code{getDistribution()}}{
Accessor function for the name of the uncertainty distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GammaModelVariable$getDistribution()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Distribution name as character string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMean"></a>}}
\if{latex}{\out{\hypertarget{method-getMean}{}}}
\subsection{Method \code{getMean()}}{
Return the expected value of the distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GammaModelVariable$getMean()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Expected value as a numeric value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSD"></a>}}
\if{latex}{\out{\hypertarget{method-getSD}{}}}
\subsection{Method \code{getSD()}}{
Return the standard deviation of the distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GammaModelVariable$getSD()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Standard deviation as a numeric value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getQuantile"></a>}}
\if{latex}{\out{\hypertarget{method-getQuantile}{}}}
\subsection{Method \code{getQuantile()}}{
Return the quantiles of the Gamma uncertainty distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GammaModelVariable$getQuantile(probs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{probs}}{Vector of probabilities, in range [0,1].}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of quantiles.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GammaModelVariable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
