% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gv2png}
\alias{gv2png}
\title{Render a DOT format graph as a png file.}
\usage{
gv2png(dot, pngfile = tempfile(fileext = ".png"))
}
\arguments{
\item{dot}{GraphViz dot representation in character vector form.}

\item{pngfile}{path of png file to create.}
}
\value{
pathname of the png file created (including extension).
}
\description{
Uses the \code{dot} command line tool from the graphviz project,
if it is available on the host system, or creates a placeholder image if not.
}
