% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovModel.R
\docType{class}
\name{MarkovModel}
\alias{MarkovModel}
\title{MarkovModel}
\description{
An R6 class for a Markov model
}
\details{
A class to represent a complete Markov model.
}
\section{Methods}{

\describe{

\item{getStatenames}{Returns a character list of state names}
\item{setPopulations}{Sets the occupancy of each state. Takes argument
        \code{populations} which is a named vector of populations for
        the start of the state. The names should be the state names. 
        Due to R's implementation of matrix algebra, \code{populations} 
        must be a numeric type and is not restricted to being an integer.
        Calling this resets the cycle count for the model to zero.}
\item{setTransitions}{Sets the annual state transition rates. Argument
        \code{Ip} must be a numeric square matrix of dimension
        equal to the number of states, with row and column names equal
        to Markov state names. Transition rates are from row state
        to column state.}
\item{stateSummary}{Creates a state summary data frame suitable for
        printing with kable etc.}
\item{transitionSummary}{Creates a table of the annual transition
        probabilities.}
}
}

\author{
Andrew J. Sims \email{andrew.sims5@nhs.net}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MarkovModel$new()}}
\item \href{#method-cycle}{\code{MarkovModel$cycle()}}
\item \href{#method-cycles}{\code{MarkovModel$cycles()}}
\item \href{#method-getStatenames}{\code{MarkovModel$getStatenames()}}
\item \href{#method-setDiscount}{\code{MarkovModel$setDiscount()}}
\item \href{#method-setPopulations}{\code{MarkovModel$setPopulations()}}
\item \href{#method-setTransitions}{\code{MarkovModel$setTransitions()}}
\item \href{#method-stateSummary}{\code{MarkovModel$stateSummary()}}
\item \href{#method-transitionSummary}{\code{MarkovModel$transitionSummary()}}
\item \href{#method-clone}{\code{MarkovModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a Markov model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovModel$new(states, Ip, discount = 0, nCyclesPerYear = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{states}}{a list of objects of type MarkovState}

\item{\code{Ip}}{Matrix of \emph{annual} rates of transition between states. The
dimensions are (nStates,nStates) with each entry being the
annual rate of transitions from the row state to the
column state. Values are in the range [0,1]. Transitions
to self (i.e. leading diagonal) should be set to zero, and will
be adjusted so that the sum of transitions from each row are
zero. The matrix should have row names and column names which are the
state names.}

\item{\code{discount}}{The annual discount rate (percentage).}

\item{\code{nCyclesPerYear}}{The number of cycles per year.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A MarkovModel object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cycle"></a>}}
\if{latex}{\out{\hypertarget{method-cycle}{}}}
\subsection{Method \code{cycle()}}{
Applies one cycle of the model, starting at zero.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovModel$cycle()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Calculated values, per state.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cycles"></a>}}
\if{latex}{\out{\hypertarget{method-cycles}{}}}
\subsection{Method \code{cycles()}}{
Applies \code{nCycles} cycles of the model. The starting
populations are redistributed through the
transition probabilities and the state occupancy costs are
calculated, using function \code{cycle}. The end populations are
then fed back into the model for a further cycle and the
process is repeated. For each cycle, the state populations and
the aggregated occupancy costs are saved in one row of the
returned dataframe, with the cycle number. If the cycle count
for the model is zero when called, the first cycle evaluated
will be cycle zero, i.e. the distribution of patients to starting
states, which will include calculation of state entry costs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovModel$cycles(nCycles = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nCycles}}{Number of cycles to run; default is 2.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame with cycle results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getStatenames"></a>}}
\if{latex}{\out{\hypertarget{method-getStatenames}{}}}
\subsection{Method \code{getStatenames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovModel$getStatenames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setDiscount"></a>}}
\if{latex}{\out{\hypertarget{method-setDiscount}{}}}
\subsection{Method \code{setDiscount()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovModel$setDiscount(r)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setPopulations"></a>}}
\if{latex}{\out{\hypertarget{method-setPopulations}{}}}
\subsection{Method \code{setPopulations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovModel$setPopulations(populations)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setTransitions"></a>}}
\if{latex}{\out{\hypertarget{method-setTransitions}{}}}
\subsection{Method \code{setTransitions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovModel$setTransitions(Ip)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stateSummary"></a>}}
\if{latex}{\out{\hypertarget{method-stateSummary}{}}}
\subsection{Method \code{stateSummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovModel$stateSummary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transitionSummary"></a>}}
\if{latex}{\out{\hypertarget{method-transitionSummary}{}}}
\subsection{Method \code{transitionSummary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovModel$transitionSummary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
