% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovState.R
\docType{class}
\name{MarkovState}
\alias{MarkovState}
\title{MarkovState}
\description{
An R6 class for a state in a Markov model
}
\details{
Class to represent a single state in a Markov model.
}
\author{
Andrew J. Sims \email{andrew.sims@newcastle.ac.uk}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MarkovState$new()}}
\item \href{#method-get_name}{\code{MarkovState$get_name()}}
\item \href{#method-has_cycle_limit}{\code{MarkovState$has_cycle_limit()}}
\item \href{#method-get_cycle_limit}{\code{MarkovState$get_cycle_limit()}}
\item \href{#method-is_absorbing}{\code{MarkovState$is_absorbing()}}
\item \href{#method-setAnnualCost}{\code{MarkovState$setAnnualCost()}}
\item \href{#method-getAnnualCost}{\code{MarkovState$getAnnualCost()}}
\item \href{#method-setEntryCost}{\code{MarkovState$setEntryCost()}}
\item \href{#method-getEntryCost}{\code{MarkovState$getEntryCost()}}
\item \href{#method-get_utility}{\code{MarkovState$get_utility()}}
\item \href{#method-clone}{\code{MarkovState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Markov state object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$new(
  name,
  annualCost = 0,
  entryCost = 0,
  utility = 1,
  cycleLimit = NA,
  absorbing = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the state (character string).}

\item{\code{annualCost}}{The annual cost of state occupancy.}

\item{\code{entryCost}}{The cost to enter the state.}

\item{\code{utility}}{The utility associated with being in the state.}

\item{\code{cycleLimit}}{The maximum length of stay;
numeric. Leave unset for no cycle limit (normal behaviour).}

\item{\code{absorbing}}{Logical; TRUE if the state is an absorbing state,
FALSE otherwise. Must be FALSE (default) if cycleLimit is set.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of type MarkovState.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-get_name}{}}}
\subsection{Method \code{get_name()}}{
Accessor function to retrieve the state name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$get_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
State name.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_cycle_limit"></a>}}
\if{latex}{\out{\hypertarget{method-has_cycle_limit}{}}}
\subsection{Method \code{has_cycle_limit()}}{
Accessor function to identify if the state has a cycle limit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$has_cycle_limit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE if temporary state; false otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_cycle_limit"></a>}}
\if{latex}{\out{\hypertarget{method-get_cycle_limit}{}}}
\subsection{Method \code{get_cycle_limit()}}{
Accessor function to find the cycle limit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$get_cycle_limit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The cycle limit; integer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_absorbing"></a>}}
\if{latex}{\out{\hypertarget{method-is_absorbing}{}}}
\subsection{Method \code{is_absorbing()}}{
Test whether this is an absorbing state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$is_absorbing()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE if state was defined as absorbing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setAnnualCost"></a>}}
\if{latex}{\out{\hypertarget{method-setAnnualCost}{}}}
\subsection{Method \code{setAnnualCost()}}{
Sets the annual cost of state occupancy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$setAnnualCost(annualCost)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{annualCost}}{Annual cost of occupying the state; numeric.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated MarkovState object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAnnualCost"></a>}}
\if{latex}{\out{\hypertarget{method-getAnnualCost}{}}}
\subsection{Method \code{getAnnualCost()}}{
Gets the annual cost of state occupancy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$getAnnualCost()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Annual cost; numeric.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setEntryCost"></a>}}
\if{latex}{\out{\hypertarget{method-setEntryCost}{}}}
\subsection{Method \code{setEntryCost()}}{
Sets the entry cost of state occupancy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$setEntryCost(entryCost)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entryCost}}{Cost to enter the state; numeric.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated MarkovState object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEntryCost"></a>}}
\if{latex}{\out{\hypertarget{method-getEntryCost}{}}}
\subsection{Method \code{getEntryCost()}}{
Gets the entry cost of state occupancy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$getEntryCost()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Entry cost; numeric.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_utility"></a>}}
\if{latex}{\out{\hypertarget{method-get_utility}{}}}
\subsection{Method \code{get_utility()}}{
Gets the utility associated with the state.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$get_utility()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Utility; numeric.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
