% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovTransition.R
\docType{class}
\name{MarkovTransition}
\alias{MarkovTransition}
\title{A transition in a Markov model}
\description{
An R6 class representing a transition in a Markov model.
}
\details{
A specialism of class \code{Arrow} which is used in a Markov model
to represent a transition between two \code{MarkovState}s. The transition
is associated with an instantaneous hazard rate and optionally, a cost. The
hazard rate is the rate of failure at the next instant given survival up to
the current time.
}
\author{
Andrew J. Sims \email{andrew.sims@newcastle.ac.uk}
}
\section{Super classes}{
\code{\link[rdecision:Edge]{rdecision::Edge}} -> \code{\link[rdecision:Arrow]{rdecision::Arrow}} -> \code{MarkovTransition}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MarkovTransition$new()}}
\item \href{#method-modvars}{\code{MarkovTransition$modvars()}}
\item \href{#method-set_rate}{\code{MarkovTransition$set_rate()}}
\item \href{#method-rate}{\code{MarkovTransition$rate()}}
\item \href{#method-cost}{\code{MarkovTransition$cost()}}
\item \href{#method-clone}{\code{MarkovTransition$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Edge" data-id="endpoints">}\href{../../rdecision/html/Edge.html#method-endpoints}{\code{rdecision::Edge$endpoints()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Edge" data-id="is_same_edge">}\href{../../rdecision/html/Edge.html#method-is_same_edge}{\code{rdecision::Edge$is_same_edge()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Edge" data-id="label">}\href{../../rdecision/html/Edge.html#method-label}{\code{rdecision::Edge$label()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Arrow" data-id="source">}\href{../../rdecision/html/Arrow.html#method-source}{\code{rdecision::Arrow$source()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Arrow" data-id="target">}\href{../../rdecision/html/Arrow.html#method-target}{\code{rdecision::Arrow$target()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an object of type \code{MarkovTransition}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovTransition$new(source, target, r = as.numeric(NA), cost = 0, label = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{\code{MarkovState} from which the transition starts.}

\item{\code{target}}{\code{MarkovState} to which the transition ends.}

\item{\code{r}}{Instantaneous hazard rate in units of per patient per year.
To meet the condition that the sum of the outgoing transition rates from
each state is zero, one transition rate from a state may be set to NA
when the model is defined. The default is NA.}

\item{\code{cost}}{Cost associated with the transition.}

\item{\code{label}}{Character string containing a label for the transition (the
name of the event).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
A hazard rate must be assigned to the transition. Optionally,
a cost may be associated with the transition.
}

\subsection{Returns}{
A new \code{MarkovTransition} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-modvars"></a>}}
\if{latex}{\out{\hypertarget{method-modvars}{}}}
\subsection{Method \code{modvars()}}{
Find all the model variables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovTransition$modvars()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Find variables of type \code{ModVar} that have been 
specified as values associated with this \code{MarkovTransition}. 
Includes operands of these \code{ModVar}s, if they are expressions.
}

\subsection{Returns}{
A list of \code{ModVar}s.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_rate"></a>}}
\if{latex}{\out{\hypertarget{method-set_rate}{}}}
\subsection{Method \code{set_rate()}}{
Set the value of the transition rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovTransition$set_rate(r = as.numeric(NA))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r}}{Instantaneous hazard rate in units of per patient per year,
or NA (the default).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated \code{MarkovTransition} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rate"></a>}}
\if{latex}{\out{\hypertarget{method-rate}{}}}
\subsection{Method \code{rate()}}{
Return the value of the hazard rate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovTransition$rate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Numeric value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cost"></a>}}
\if{latex}{\out{\hypertarget{method-cost}{}}}
\subsection{Method \code{cost()}}{
Return the cost associated with traversing the edge.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovTransition$cost()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Cost.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovTransition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
