% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelVariable.R
\docType{class}
\name{ModelVariable}
\alias{ModelVariable}
\title{ModelVariable}
\description{
An R6 class for a variable in an health economic model
}
\details{
Base class for a variable used in a health economic model. The base 
class, which is not intended to be directly instantiated by model
applications, wraps a numerical value which is used in calculations.
The base class provides a framework for creating classes of model
variables whose uncertainties are described by statistical distributions
parametrized with hyperparameters.
}
\author{
Andrew Sims \email{andrew.sims5@nhs.net}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelVariable$new()}}
\item \href{#method-sample}{\code{ModelVariable$sample()}}
\item \href{#method-value}{\code{ModelVariable$value()}}
\item \href{#method-getDescription}{\code{ModelVariable$getDescription()}}
\item \href{#method-getUnits}{\code{ModelVariable$getUnits()}}
\item \href{#method-getDistribution}{\code{ModelVariable$getDistribution()}}
\item \href{#method-getMean}{\code{ModelVariable$getMean()}}
\item \href{#method-getSD}{\code{ModelVariable$getSD()}}
\item \href{#method-getQuantile}{\code{ModelVariable$getQuantile()}}
\item \href{#method-clone}{\code{ModelVariable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an object of type `ModelVariable`
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelVariable$new(description, units)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{A character string description of the variable
and its role in the
model. This description will be used in a tabulation of the
variables linked to a model.}

\item{\code{units}}{A character string description of the units, e.g. 'GBP',
'per year'.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new ModelVariable object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
Set the value of the model variable from its uncertainty distribution.
Nothing is returned; the sampled value is returned at the next
call to `value()`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelVariable$sample(expected = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expected}}{Logical; if TRUE sets the value of the model variable
returned at subsequent calls to `value()` to be equal to the 
expectation of the variable. Default is FALSE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated ModelVariable object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-value"></a>}}
\if{latex}{\out{\hypertarget{method-value}{}}}
\subsection{Method \code{value()}}{
Return the current value of the model variable. This will be the 
expected value if the argument to the most recent call to `sample`
was TRUE or after creation of the object; otherwise it will return
a value sampled from the uncertainty distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelVariable$value()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Numeric value of the model variable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDescription"></a>}}
\if{latex}{\out{\hypertarget{method-getDescription}{}}}
\subsection{Method \code{getDescription()}}{
Accessor functions for the description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelVariable$getDescription()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Description of model variable as character string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getUnits"></a>}}
\if{latex}{\out{\hypertarget{method-getUnits}{}}}
\subsection{Method \code{getUnits()}}{
Accessor function for units.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelVariable$getUnits()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Description of units as character string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDistribution"></a>}}
\if{latex}{\out{\hypertarget{method-getDistribution}{}}}
\subsection{Method \code{getDistribution()}}{
Accessor function for the name of the uncertainty distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelVariable$getDistribution()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Distribution name as character string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMean"></a>}}
\if{latex}{\out{\hypertarget{method-getMean}{}}}
\subsection{Method \code{getMean()}}{
Return the expected value of the distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelVariable$getMean()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Expected value as a numeric value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSD"></a>}}
\if{latex}{\out{\hypertarget{method-getSD}{}}}
\subsection{Method \code{getSD()}}{
Return the standard deviation of the distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelVariable$getSD()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Standard deviation as a numeric value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getQuantile"></a>}}
\if{latex}{\out{\hypertarget{method-getQuantile}{}}}
\subsection{Method \code{getQuantile()}}{
Find quantiles of the uncertainty distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelVariable$getQuantile(probs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{probs}}{Numeric vector of probabilities, each in range [0,1].}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of numeric values of the same length as `probs`.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelVariable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
