% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelVariableExpression.R
\docType{class}
\name{ModelVariableExpression}
\alias{ModelVariableExpression}
\title{ModelVariableExpression}
\description{
An R6 class for an expression involving model variables
}
\details{
A class to support expressions involving objects
of base class ModelVariable. For example if \code{A} and
\code{B} are variables with base class \code{ModelVariable}
and \code{c} is a variable of type \code{numeric}, then
it is not possible to write, for example, 
\code{x <- 42*A/B + c}, because R cannot manipulate class
variables using the same operators as regular variables. 
But such forms of expression may be desirable in constructing a
model and this class provides a mechanism for doing so.
Instances of the class should be created by providing
an expression involving model variables, and evaluated
by making a call to \code{eval}. 
In its simplest form, if \code{A} is a model variable, then
quote(A) is a ModelVariableExpression.
}
\note{
TO DO: does not work if expr is not in the global namespace.
}
\author{
Andrew J. Sims \email{andrew.sims5@nhs.net}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelVariableExpression$new()}}
\item \href{#method-eval}{\code{ModelVariableExpression$eval()}}
\item \href{#method-getModelVariables}{\code{ModelVariableExpression$getModelVariables()}}
\item \href{#method-clone}{\code{ModelVariableExpression$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a model variable expression
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelVariableExpression$new(expr, description = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expr}}{An R expression involving model variables which would be 
syntactically correct were each model variable to be replaced by
numerical variables.}

\item{\code{description}}{Label for the model variable expresssion. In 
a complex model it may help to tabulate how model variables are
combined into costs, probablities and rates.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of type ModelVariableExpression
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval"></a>}}
\if{latex}{\out{\hypertarget{method-eval}{}}}
\subsection{Method \code{eval()}}{
Evaluate the expression.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelVariableExpression$eval(expected = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expected}}{If TRUE, evaluate each model variable at its expectation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Numerical value of the evaluated expression.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getModelVariables"></a>}}
\if{latex}{\out{\hypertarget{method-getModelVariables}{}}}
\subsection{Method \code{getModelVariables()}}{
Return a list of named ModelVariables associated with the expression.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelVariableExpression$getModelVariables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list of model variables. Each member of the list is
named with a character string of the variable name used in constructing
the model. This is not necessary for evaluation, but helps when tabulating
the variables used in a model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelVariableExpression$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
