% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModVar.R
\docType{class}
\name{ModVar}
\alias{ModVar}
\title{A model variable incorporating uncertainty}
\description{
An R6 class for a variable in a health economic model.
}
\details{
Base class for a variable used in a health economic model. The base 
class wraps a numerical value which is used in calculations.
It provides a framework for creating classes of model
variables whose uncertainties are described by statistical distributions
parametrized with hyperparameters.
}
\author{
Andrew Sims \email{andrew.sims@newcastle.ac.uk}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModVar$new()}}
\item \href{#method-is_expression}{\code{ModVar$is_expression()}}
\item \href{#method-is_probabilistic}{\code{ModVar$is_probabilistic()}}
\item \href{#method-description}{\code{ModVar$description()}}
\item \href{#method-units}{\code{ModVar$units()}}
\item \href{#method-distribution}{\code{ModVar$distribution()}}
\item \href{#method-mean}{\code{ModVar$mean()}}
\item \href{#method-mode}{\code{ModVar$mode()}}
\item \href{#method-SD}{\code{ModVar$SD()}}
\item \href{#method-quantile}{\code{ModVar$quantile()}}
\item \href{#method-r}{\code{ModVar$r()}}
\item \href{#method-set}{\code{ModVar$set()}}
\item \href{#method-get}{\code{ModVar$get()}}
\item \href{#method-clone}{\code{ModVar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an object of type \code{ModVar}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModVar$new(description, units, D = NULL, k = as.integer(1))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{A character string description of the variable
and its role in the model. This description will be used in a
tabulation of the variables linked to a model.}

\item{\code{units}}{A character string description of the units, e.g. 
\code{"GBP"}, \code{"per year"}.}

\item{\code{D}}{The distribution representing the uncertainty in the variable.
Should inherit from class \code{Distribution}, or NULL if none is
defined.}

\item{\code{k}}{The index of the dimension of the multivariate distribution
that applies to this model variable.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
A \code{ModVar} is associated with an uncertainty distribution
(a "has-a" relationship in object-oriented terminology). There can be a
1-1 mapping of \code{ModVar}s to \code{Distribution}s, or several
model variables can be linked to the same distribution in a
many-1 mapping, e.g. when each transition probability from a Markov state
is represented as a \code{ModVar} and each can be linked to the \code{k}
dimensions of a common multivariate Dirichlet distribution.
}

\subsection{Returns}{
A new \verb{ModVar} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_expression"></a>}}
\if{latex}{\out{\hypertarget{method-is_expression}{}}}
\subsection{Method \code{is_expression()}}{
Is this \code{ModVar} an expression?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModVar$is_expression()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if it inherits from \code{ExprModVar}, \code{FALSE}
otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_probabilistic"></a>}}
\if{latex}{\out{\hypertarget{method-is_probabilistic}{}}}
\subsection{Method \code{is_probabilistic()}}{
Is the model variable probabilistic?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModVar$is_probabilistic()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Tests whether the model variable is probabilistic, i.e. a 
random variable that follows a distribution, or an expression involving
random variables, some of which follow distributions.
}

\subsection{Returns}{
\code{TRUE} if probabilistic
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-description"></a>}}
\if{latex}{\out{\hypertarget{method-description}{}}}
\subsection{Method \code{description()}}{
Accessor function for the description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModVar$description()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Description of model variable as character string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-units"></a>}}
\if{latex}{\out{\hypertarget{method-units}{}}}
\subsection{Method \code{units()}}{
Accessor function for units.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModVar$units()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Description of units as character string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distribution"></a>}}
\if{latex}{\out{\hypertarget{method-distribution}{}}}
\subsection{Method \code{distribution()}}{
Name and parameters of the uncertainty distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModVar$distribution()}\if{html}{\out{</div>}}
}

\subsection{Details}{
If \eqn{K > 1} the dimension of the distribution associated 
with this model variable is appended, e.g. \code{Dir(2,3)[1]} 
means that the model variable is associated with the first dimension 
of a 2D Dirichlet distribution with alpha parameters 2 and 3.
}

\subsection{Returns}{
Distribution name as character string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mean"></a>}}
\if{latex}{\out{\hypertarget{method-mean}{}}}
\subsection{Method \code{mean()}}{
Mean value of the model variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModVar$mean()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Mean value as a numeric value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mode"></a>}}
\if{latex}{\out{\hypertarget{method-mode}{}}}
\subsection{Method \code{mode()}}{
The mode of the variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModVar$mode()}\if{html}{\out{</div>}}
}

\subsection{Details}{
By default returns \code{NA}, which will be the case for 
most \code{ModVar} variables, because arbitrary distributions are 
not guaranteed to be unimodal.
}

\subsection{Returns}{
Mode as a numeric value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SD"></a>}}
\if{latex}{\out{\hypertarget{method-SD}{}}}
\subsection{Method \code{SD()}}{
Standard deviation of the model variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModVar$SD()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Standard deviation as a numeric value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-quantile"></a>}}
\if{latex}{\out{\hypertarget{method-quantile}{}}}
\subsection{Method \code{quantile()}}{
Quantiles of the uncertainty distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModVar$quantile(probs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{probs}}{Numeric vector of probabilities, each in range [0,1].}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of numeric values of the same length as \code{probs}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-r"></a>}}
\if{latex}{\out{\hypertarget{method-r}{}}}
\subsection{Method \code{r()}}{
Draw a random sample from the model variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModVar$r()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The same random sample will be returned until \code{set} is
called to force a resample.
}

\subsection{Returns}{
A sample drawn at random.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
Sets the value of the \code{ModVar}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModVar$set(what = "random", val = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{what}}{Character: one of \code{"random"} (samples from the 
uncertainty distribution), \code{"expected"} (mean), \code{"q2.5"}
(lower 95\% confidence limit), \code{"q50"} (median), \code{"q97.5"}
(upper 95\% confidence limit), \code{"current"} (leaves
the value unchanged), \code{"value"} (sets the value explicitly).}

\item{\code{val}}{A numeric value, only used with \code{what}=\code{"value"}, 
ignored otherwise.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Defines what will be returned by subsequent
calls to \code{get()} until \code{set()} is called again. The 
\code{"current"} option is provided to support having common 
functions to set (or leave alone) sets of model variables, depending on
their use case and avoids additional if statements. Option \code{"value"}
is not recommended for normal usage because it allows the model variable
to be set to an implausible value, based on its defined uncertainty. An 
example of where this may be needed is in threshold finding.
}

\subsection{Returns}{
Updated \code{ModVar}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
Get the value of the \code{ModVar}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModVar$get()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns the value defined by the most recent call
to \code{set()}.
}

\subsection{Returns}{
Value determined by last \code{set()}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModVar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
