% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.apply.R
\name{node.apply}
\alias{node.apply}
\title{An apply function for all nodes in a decision tree}
\usage{
node.apply(node, FUN, ...)
}
\arguments{
\item{node}{first node in the tree to which FUN should be applied.}

\item{FUN}{function to apply to each node in the subtree.}

\item{...}{optional arguments to FUN.}
}
\value{
A list of length equal to the number of nodes in the 
        tree with each member of the list equal to the result
        of applying FUN to that node. For example,
        if FUN=length, the return value will be the number of
        nodes in the tree.
}
\description{
\code{node.apply} traverses each node path
             in the tree, starting with \code{node}, and applies
             function FUN to each node in the (sub-)tree.
}
\author{
Andrew J. Sims \email{andrew.sims5@nhs.net}
}
