% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.mvlist.R
\name{node.mvlist}
\alias{node.mvlist}
\title{Returns a list of model variables associated with a node}
\usage{
node.mvlist(nodes)
}
\arguments{
\item{nodes}{List of objects of type 'Node'.}
}
\value{
Unique list of model variables associated with any part of the subtree.
}
\description{
Each node may have several ModelVariables associated with
it, directly as costs, utilities etc, or indirectly via 
ModelVariableExpressions that have been supplied as costs, probabilities, 
utilities etc. This function is intended for use with \code{node.apply}
which traverses each node in a (sub-)tree.
}
\author{
Andrew J. Sims \email{andrew.sims5@nhs.net}
}
