% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Node.R
\docType{class}
\name{Node}
\alias{Node}
\title{Node}
\description{
An R6 class to represent a node in a decision tree
}
\details{
Base class to represent a single node in a decision tree. Non subclassed
nodes are not expected to be created as model objects. Document Object
Model (DOM) names are used for node methods as far as possibe.
}
\author{
Andrew Sims \email{andrew.sims5@nhs.net}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Node$new()}}
\item \href{#method-hasChildNodes}{\code{Node$hasChildNodes()}}
\item \href{#method-childNodes}{\code{Node$childNodes()}}
\item \href{#method-descendantNodes}{\code{Node$descendantNodes()}}
\item \href{#method-isSameNode}{\code{Node$isSameNode()}}
\item \href{#method-nodeType}{\code{Node$nodeType()}}
\item \href{#method-getPathways}{\code{Node$getPathways()}}
\item \href{#method-getLabel}{\code{Node$getLabel()}}
\item \href{#method-getP}{\code{Node$getP()}}
\item \href{#method-getUtility}{\code{Node$getUtility()}}
\item \href{#method-getCost}{\code{Node$getCost()}}
\item \href{#method-get_modvars}{\code{Node$get_modvars()}}
\item \href{#method-tabulate_modvars}{\code{Node$tabulate_modvars()}}
\item \href{#method-sample_modvars}{\code{Node$sample_modvars()}}
\item \href{#method-updateEdges}{\code{Node$updateEdges()}}
\item \href{#method-clone}{\code{Node$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new Node object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new Node object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hasChildNodes"></a>}}
\if{latex}{\out{\hypertarget{method-hasChildNodes}{}}}
\subsection{Method \code{hasChildNodes()}}{
Does the node have any child nodes? (DOM-style)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$hasChildNodes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE if node has children, FALSE if not
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-childNodes"></a>}}
\if{latex}{\out{\hypertarget{method-childNodes}{}}}
\subsection{Method \code{childNodes()}}{
Return list of child nodes (DOM-style)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$childNodes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of child Nodes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-descendantNodes"></a>}}
\if{latex}{\out{\hypertarget{method-descendantNodes}{}}}
\subsection{Method \code{descendantNodes()}}{
Return list of descendent nodes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$descendantNodes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of descendent nodes, including self.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isSameNode"></a>}}
\if{latex}{\out{\hypertarget{method-isSameNode}{}}}
\subsection{Method \code{isSameNode()}}{
Is this node the same as the argument? (DOM-style)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$isSameNode(otherNode)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{otherNode}}{node to compare with this one}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if `otherNode` is also this one
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodeType"></a>}}
\if{latex}{\out{\hypertarget{method-nodeType}{}}}
\subsection{Method \code{nodeType()}}{
node type (DOM-style)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$nodeType()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Node class, as character string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPathways"></a>}}
\if{latex}{\out{\hypertarget{method-getPathways}{}}}
\subsection{Method \code{getPathways()}}{
Trace and list all pathways ending on leaf nodes which start
with this node.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$getPathways(choice = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{choice}}{Name of choice. All pathways are returned if NA.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of Path objects. Each member of the list is a
path from this node to a leaf, limited to those associated
with choice, if defined.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLabel"></a>}}
\if{latex}{\out{\hypertarget{method-getLabel}{}}}
\subsection{Method \code{getLabel()}}{
Return label of edge which links to specified child node
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$getLabel(childNode)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{childNode}}{child node to which find label of linking edge}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
label as character string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getP"></a>}}
\if{latex}{\out{\hypertarget{method-getP}{}}}
\subsection{Method \code{getP()}}{
Function to return the conditional probability of the edge which links to
the specified child node
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$getP(childNode)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{childNode}}{child node to which to find probability of linking edge}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
numerical value of probability
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getUtility"></a>}}
\if{latex}{\out{\hypertarget{method-getUtility}{}}}
\subsection{Method \code{getUtility()}}{
function to return the utility associated with the node
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$getUtility()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Utility, numeric
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCost"></a>}}
\if{latex}{\out{\hypertarget{method-getCost}{}}}
\subsection{Method \code{getCost()}}{
Function to return the cost of the edge which links to the specified child node
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$getCost(childNode)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{childNode}}{child node to identify edge with associated cost of traversal}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Cost, numerical value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_modvars"></a>}}
\if{latex}{\out{\hypertarget{method-get_modvars}{}}}
\subsection{Method \code{get_modvars()}}{
Function to return a list of model variables associated with this node.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$get_modvars()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of model variables associated with this node.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tabulate_modvars"></a>}}
\if{latex}{\out{\hypertarget{method-tabulate_modvars}{}}}
\subsection{Method \code{tabulate_modvars()}}{
Tabulate all model variables associated with this node.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$tabulate_modvars(include.descendants = FALSE, include.operands = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include.descendants}}{If TRUE, model variables associated
with this node and its descendants are tabulated; otherwise only
the ones that are associated with this node.}

\item{\code{include.operands}}{If TRUE, recursively add model variables which are
included in expressions in ExprModVars. Default is
FALSE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame with one row per model variable, as follows:
\describe{
\item{Label}{The label given to the variable on creation.}
\item{Description}{As given at initialization.}
\item{Units}{Units of the variable.}
\item{Distribution}{The uncertainty distribution or an expression.} 
\item{Mean}{Expected value.}
\item{SD}{Standard deviation.}
\item{Q2.5}{p=0.025 quantile.}
\item{Q97.5}{p=0.975 quantile.}
\item{Qhat}{Asterisk if the quantiles and SD were estimated by random sampling.}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_modvars"></a>}}
\if{latex}{\out{\hypertarget{method-sample_modvars}{}}}
\subsection{Method \code{sample_modvars()}}{
Sample the model variables associated with the node.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$sample_modvars(expected = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expected}}{if TRUE, use the expected value of the model variables in
the node; otherwise sample from their uncertainty distributions.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated Node object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateEdges"></a>}}
\if{latex}{\out{\hypertarget{method-updateEdges}{}}}
\subsection{Method \code{updateEdges()}}{
Update the values on the edges associated with the node.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$updateEdges()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Updated Node object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
