% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalModelVariable.R
\docType{class}
\name{NormalModelVariable}
\alias{NormalModelVariable}
\title{NormalModelVariable}
\description{
An R6 class for a model variable with Normal uncertainty
}
\details{
A model variable for which the uncertainty in the point estimate can
be modelled with a Normal distribution. The hyperparameters of the
distribution are the mean (`mu`) and the standard deviation (`sd`) of
the uncertainty distribution. The value of `mu` is the expected value
of the variable.
}
\author{
Andrew J. Sims \email{andrew.sims5@nhs.net}
}
\section{Super class}{
\code{\link[rdecision:ModelVariable]{rdecision::ModelVariable}} -> \code{NormalModelVariable}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NormalModelVariable$new()}}
\item \href{#method-sample}{\code{NormalModelVariable$sample()}}
\item \href{#method-getDistribution}{\code{NormalModelVariable$getDistribution()}}
\item \href{#method-getMean}{\code{NormalModelVariable$getMean()}}
\item \href{#method-getSD}{\code{NormalModelVariable$getSD()}}
\item \href{#method-clone}{\code{NormalModelVariable$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModelVariable" data-id="getDescription">}\href{../../rdecision/html/ModelVariable.html#method-getDescription}{\code{rdecision::ModelVariable$getDescription()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModelVariable" data-id="getQuantile">}\href{../../rdecision/html/ModelVariable.html#method-getQuantile}{\code{rdecision::ModelVariable$getQuantile()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModelVariable" data-id="getUnits">}\href{../../rdecision/html/ModelVariable.html#method-getUnits}{\code{rdecision::ModelVariable$getUnits()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="ModelVariable" data-id="value">}\href{../../rdecision/html/ModelVariable.html#method-value}{\code{rdecision::ModelVariable$value()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a model variable with normal uncertainty.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalModelVariable$new(description, units, mu, sigma)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{description}}{A character string describing the variable.}

\item{\code{units}}{Units of the quantity; character string.}

\item{\code{mu}}{Hyperparameter with mean of the Normal distribution for 
the uncertainty of the variable.}

\item{\code{sigma}}{Hyperparameter equal to the standard deviation of the
normal distribution for the uncertainty of the variable.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A NormalModelVariable object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
Set the value of the model variable from its uncertainty distribution.
Nothing is returned; the sampled value is returned at the next
call to `value()`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalModelVariable$sample(expected = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expected}}{Logical; if TRUE sets the value of the model variable
returned at subsequent calls to `value()` to be equal to the 
expectation of the variable. Default is FALSE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated NormalModelVariable object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDistribution"></a>}}
\if{latex}{\out{\hypertarget{method-getDistribution}{}}}
\subsection{Method \code{getDistribution()}}{
Accessor function for the name of the uncertainty distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalModelVariable$getDistribution()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Distribution name as character string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMean"></a>}}
\if{latex}{\out{\hypertarget{method-getMean}{}}}
\subsection{Method \code{getMean()}}{
Return the expected value of the distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalModelVariable$getMean()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Expected value as a numeric value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSD"></a>}}
\if{latex}{\out{\hypertarget{method-getSD}{}}}
\subsection{Method \code{getSD()}}{
Return the standard deviation of the distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalModelVariable$getSD()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Standard deviation as a numeric value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalModelVariable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
