% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.apply.R
\name{path.apply}
\alias{path.apply}
\title{An apply function for node-to-leaf paths in a decision tree.}
\usage{
path.apply(node, FUN, ...)
}
\arguments{
\item{node}{first node in the tree to which FUN should be applied.}

\item{FUN}{function to apply to each list of nodes starting with
\code{node} and ending with one of the leaf nodes. The
first agument to FUN must be a list of Nodes.}

\item{...}{optional arguments to FUN.}
}
\value{
A list of length equal to the number of node-to-leaf paths
        in the tree with each member of the list equal to the result
        of applying FUN to that list of node-to-leaf nodes. For example,
        if FUN=length, the return value will be a list of the node
        count of each node-to-leaf traversal of the tree.
}
\description{
\code{path.apply} traverses each node-to-leaf path
             in the tree, starting with @code{node}, and applies
             function FUN to each list of nodes in each node-to-leaf
             path.
}
