% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path.R
\docType{class}
\name{Path}
\alias{Path}
\title{Path}
\description{
An R6 class for pathway through a decision tree.
}
\details{
A class to represent a traversal through a decision tree. Essentially
a wrapper for an ordered list of Nodes.
}
\note{
Uses the first decision node in the list.

Assumes the leaf node is the last node in the list.

Assumes the nodes are supplied in order of traversal and that
the final node is a leaf node.

Assumes that the final node is a leaf node.

Assumes that only the final leaf node has a utility value.
}
\author{
Andrew J. Sims \email{andrew.sims5@nhs.net}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Path$new()}}
\item \href{#method-getChoice}{\code{Path$getChoice()}}
\item \href{#method-getName}{\code{Path$getName()}}
\item \href{#method-getProbability}{\code{Path$getProbability()}}
\item \href{#method-getCost}{\code{Path$getCost()}}
\item \href{#method-getUtility}{\code{Path$getUtility()}}
\item \href{#method-tabulate}{\code{Path$tabulate()}}
\item \href{#method-clone}{\code{Path$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a path from a list of nodes. Must end on a leaf node.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Path$new(nodes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nodes}}{A list of Nodes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getChoice"></a>}}
\if{latex}{\out{\hypertarget{method-getChoice}{}}}
\subsection{Method \code{getChoice()}}{
Returns the name of the choice arising from the first decision
node of a list of nodes. In many decision trees, where the decision 
node is at the left, many leaf nodes
(pathway names) will be associated with the same choice.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Path$getChoice()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
label of the choice; specifically the \code{edgelabel}
        field of the first decision node in \code{nodes}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getName"></a>}}
\if{latex}{\out{\hypertarget{method-getName}{}}}
\subsection{Method \code{getName()}}{
Returns the name of the pathway from the (final) leaf
node of a list of nodes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Path$getName()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
name of pathway; specifically the @code{pathway}
        field of the final leaf node in @code{nodes}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProbability"></a>}}
\if{latex}{\out{\hypertarget{method-getProbability}{}}}
\subsection{Method \code{getProbability()}}{
Calculates the product of the conditional P values in
the tree traversal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Path$getProbability()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Product of conditional probabilities.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCost"></a>}}
\if{latex}{\out{\hypertarget{method-getCost}{}}}
\subsection{Method \code{getCost()}}{
Calculates the sum of the pathway costs in
the tree traversal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Path$getCost()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Sum of the pathway costs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getUtility"></a>}}
\if{latex}{\out{\hypertarget{method-getUtility}{}}}
\subsection{Method \code{getUtility()}}{
Calculates the sum of the pathway utilities in
the tree traversal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Path$getUtility()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Sum of pathway utilities
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tabulate"></a>}}
\if{latex}{\out{\hypertarget{method-tabulate}{}}}
\subsection{Method \code{tabulate()}}{
Evaluate and tabulate the pathway.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Path$tabulate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data frame with one row, as follows:
\describe{
\item{Choice}{The choice from which the pathway leads.}
\item{Pathway}{The end-point or outcome; i.e. label ofleaf node.}
\item{Probability}{Product of probabilities.}
\item{Cost}{Sum of cost.}
\item{Utility}{Sum of utility.}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Path$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
