% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.choice.R
\name{pathway.choice}
\alias{pathway.choice}
\title{Returns the choice label associated with the first decision node
in a list of nodes.}
\usage{
pathway.choice(nodes)
}
\arguments{
\item{nodes}{list of objects of type 'Node'.}
}
\value{
label of the choice; specifically the \code{edgelabel}
        field of the first decision node in \code{nodes}.
}
\description{
Returns the name of the choice arising from the first decision
node of a list of nodes. Intended to be used with \code{path.apply}
to return a list of choices for each node-to-leaf traversal. In many
decision trees, where the decision node is at the left, many leaf nodes
(pathway names) will be associated with the same choice.
}
\note{
Uses the first decision node in the list.
}
