% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.cost.R
\name{pathway.cost}
\alias{pathway.cost}
\title{Calculate pathway cost from decision and chance nodes.}
\usage{
pathway.cost(nodes)
}
\arguments{
\item{nodes}{List of nodes in the node-to-leaf traversal path.}
}
\description{
Calculates the sum of the pathways costs in
the tree traversal.
}
\note{
Assumes that the final node is a leaf node.
}
