% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.name.R
\name{pathway.name}
\alias{pathway.name}
\title{Return leaf name from a leaf node in a  list of nodes.}
\usage{
pathway.name(nodes)
}
\arguments{
\item{nodes}{list of objects of type 'Node'.}
}
\value{
name of pathway; specifically the @code{pathway}
        field of the final leaf node in @code{nodes}.
}
\description{
Returns the name of the pathway from the (final) leaf
node of a list of nodes. Intended to be used with @code{path.apply}
to return a list of leaf node names for each node-to-leaf traversal.
}
\note{
Assumes the leaf node is the last node in the list.
}
