% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.probability.R
\name{pathway.probability}
\alias{pathway.probability}
\title{Calculate pathway probability from conditional probabilities.}
\usage{
pathway.probability(nodes)
}
\arguments{
\item{nodes}{List of nodes in the node-to-leaf traversal path.}
}
\description{
Calculates the product of the conditional P values in
the tree traversal.
}
\note{
Assumes the nodes are supplied in order of traversal and that
the final node is a leaf node.
}
