% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.utility.R
\name{pathway.utility}
\alias{pathway.utility}
\title{Calculate pathway utility.}
\usage{
pathway.utility(nodes)
}
\arguments{
\item{nodes}{List of nodes in the node-to-leaf traversal path.}
}
\description{
Calculates the sum of the pathway utilities in
the tree traversal.
}
\note{
Assumes that only the final leaf node has a utility value.
}
