% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdecision-package.R
\docType{package}
\name{rdecision-package}
\alias{rdecision}
\alias{rdecision-package}
\title{rdecision: Decision Analytic Modelling In Health Economics}
\description{
A package for decision analytic modelling in health economics.
  It includes functions for modelling healthcare interventions using
  cohort models (decision trees, Markov models and extended Markov models).
  It draws on terminology from Briggs, Claxton and Sculpher, "Decision
  Modelling for Health Economic Evaluation", Oxford University Press, 2006.
}
\author{
\strong{Maintainer}: Andrew Sims \email{andrew.sims@newcastle.ac.uk}

Authors:
\itemize{
  \item Kim Fairbairn \email{kim.fairbairn@nhs.net}
}

}
\keyword{internal}
