% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDecision.R
\docType{package}
\name{rDecision}
\alias{rDecision}
\alias{rDecision-package}
\title{rDecision}
\description{
It determines best decision based on different criterias: Laplace, optimistic, pesimistc, pesimistic, Hurwicz and Savage.
}
\author{
Ruben Escribano-Garcia <escribano.enginer@gmail.com>
}
