% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tornado_plot}
\alias{tornado_plot}
\title{Draw a tornado diagram.}
\usage{
tornado_plot(to, outcome_mean, xlab = "")
}
\arguments{
\item{to}{A data frame with one row per horizontal bar of the tornado plot,
with columns:
\describe{
  \item{Description}{Description of the variable.}
  \item{Units}{Units of the variable.}
  \item{LL}{Lower limit of the variable.}
  \item{UL}{Upper limit of the variable.}
  \item{outcome.min}{Minimum value of the outcome measure.}
  \item{outcome.max}{Maximum value of the outcome measure.}
}}

\item{outcome_mean}{Mean (base case) outcome.}

\item{xlab}{Label for x axis}
}
\description{
A function to plot a tornado diagram, given a data frame of
model variable names, their confidence limits and the values of the outcome
variable for each. The outcome measure (x axis) is expected to be cost, ICER,
net monetary benefit, etc., but can be any meaningful quantity.
}
