% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionMatrix.R
\docType{class}
\name{TransitionMatrix}
\alias{TransitionMatrix}
\title{TransitionMatrix}
\description{
An R6 class for a transition matrix in Markov model
}
\details{
A class to represent a matrix of annual transition rates in a Markov
model. It wraps a regular numeric matrix, and provides methods for 
convenience and runtime checks. States are referred to by name, rather 
than by index name. Name them well!
}
\author{
Andrew J. Sims \email{andrew.sims@newcastle.ac.uk}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TransitionMatrix$new()}}
\item \href{#method-nrow}{\code{TransitionMatrix$nrow()}}
\item \href{#method-ncol}{\code{TransitionMatrix$ncol()}}
\item \href{#method-dimnames}{\code{TransitionMatrix$dimnames()}}
\item \href{#method-set_rate}{\code{TransitionMatrix$set_rate()}}
\item \href{#method-value}{\code{TransitionMatrix$value()}}
\item \href{#method-clone}{\code{TransitionMatrix$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new state transition matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$new(statenames)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{statenames}}{A character vector of state names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new TransitionMatrix object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nrow"></a>}}
\if{latex}{\out{\hypertarget{method-nrow}{}}}
\subsection{Method \code{nrow()}}{
Return the number of rows in the matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$nrow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Row count.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ncol"></a>}}
\if{latex}{\out{\hypertarget{method-ncol}{}}}
\subsection{Method \code{ncol()}}{
Return the number of columns in the matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$ncol()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Row count.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dimnames"></a>}}
\if{latex}{\out{\hypertarget{method-dimnames}{}}}
\subsection{Method \code{dimnames()}}{
Return list of state names (as per matrix dimnames).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$dimnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Named list of length 2 ('from' and 'to').
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_rate"></a>}}
\if{latex}{\out{\hypertarget{method-set_rate}{}}}
\subsection{Method \code{set_rate()}}{
Set the annual transition rate between two states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$set_rate(from, to, rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from}}{Name of the 'from' state.}

\item{\code{to}}{Name of the 'to' state.}

\item{\code{rate}}{Annual transition rate between the states. Must be a numeric
value, ModVar or NA. The latter is used to set the cell in the row which
will be computed when value() is called, to ensure that the sum of 
probabilities leaving a row is one.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated TransitionMatrix object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-value"></a>}}
\if{latex}{\out{\hypertarget{method-value}{}}}
\subsection{Method \code{value()}}{
Calculate and return the current numeric representation of the 
transition matrix. If the transition matrix was defined with one or
more transitions as model variables, the current value of these
is retrieved using 'value()'. The single NA in each row is replaced
with 1-p where P is the sum of the other values in the row.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$value()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric matrix with rows and columns labelled with 
state names. Satisfies 'is.matrix()' test in R.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransitionMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
