% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChanceNode.R
\docType{class}
\name{ChanceNode}
\alias{ChanceNode}
\title{ChanceNode}
\description{
An R6 class to represent a chance node in a decision tree.
}
\details{
An R6 class to represent a chance node in a decision tree. 
The node is associated with at least two branches to other nodes, each 
of which has a conditional probability (the probability of following
that branch given that the node has been reached) and a cost.
}
\author{
Andrew Sims \email{andrew.sims@newcastle.ac.uk}
}
\section{Super class}{
\code{\link[rdecision:Node]{rdecision::Node}} -> \code{ChanceNode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ChanceNode$new()}}
\item \href{#method-get_p}{\code{ChanceNode$get_p()}}
\item \href{#method-get_modvars}{\code{ChanceNode$get_modvars()}}
\item \href{#method-clone}{\code{ChanceNode$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="child_nodes">}\href{../../rdecision/html/Node.html#method-child_nodes}{\code{rdecision::Node$child_nodes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="descendantNodes">}\href{../../rdecision/html/Node.html#method-descendantNodes}{\code{rdecision::Node$descendantNodes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="get_edge_label">}\href{../../rdecision/html/Node.html#method-get_edge_label}{\code{rdecision::Node$get_edge_label()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="get_label">}\href{../../rdecision/html/Node.html#method-get_label}{\code{rdecision::Node$get_label()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="getPathways">}\href{../../rdecision/html/Node.html#method-getPathways}{\code{rdecision::Node$getPathways()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="has_child_nodes">}\href{../../rdecision/html/Node.html#method-has_child_nodes}{\code{rdecision::Node$has_child_nodes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="is_same_node">}\href{../../rdecision/html/Node.html#method-is_same_node}{\code{rdecision::Node$is_same_node()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="node_type">}\href{../../rdecision/html/Node.html#method-node_type}{\code{rdecision::Node$node_type()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="sample_modvars">}\href{../../rdecision/html/Node.html#method-sample_modvars}{\code{rdecision::Node$sample_modvars()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="tabulate_modvars">}\href{../../rdecision/html/Node.html#method-tabulate_modvars}{\code{rdecision::Node$tabulate_modvars()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new ChanceNode object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChanceNode$new(children, edgelabels, p, ptype = "auto")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{children}}{a list of \eqn{k \ge 2} Nodes which will be the children
of this node.}

\item{\code{edgelabels}}{a list of \eqn{k} character strings to label each edge 
(branch) leaving the `ChanceNode`, given in the same order as
`children`.}

\item{\code{p}}{a list of probabilities associated with \eqn{k} branches.
       There are four possible configurations of the list 
       based on the \code{ptype} argument.
\describe{
\item{'numeric'}{\eqn{k} numeric values; the simplest case in which the   
probabilities are certain. Supplied values should add
to unity and be given in the same order as `children`.}
\item{'MV'}{At least one `ModVar` and exactly one
`NA``, and the remainder either `numeric` or `ModVar`, given in the
same order as `children`. The single NA will be replaced on evaluation of
the model variables by a value to ensure the sum of probabilities
is unity. This option is not recommended, because there is
a chance that during sampling, individual branch probabilities
may be less than zero, or that the sum of the sampled model variable
expressions may exceed 1.}
\item{'Beta'}{One `BetaModVar` and one NA. Used for \eqn{k=2}. The 
element defined as NA will be replaced by one minus the sampled 
value of the supplied beta distribution.}
\item{'Dirichlet'}{One DirichletModVar, with \eqn{k} parameters,
given in the same order as `children`.}
\item{'auto'}{Infer which of the previous options applies based on
the types of elements of p}
}}

\item{\code{ptype}}{a character string taking one of four possible
values to define how the \code{p} argument is defined:
'numeric', 'MV', 'Beta' or 'Dirichlet'.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ChanceNode` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_p"></a>}}
\if{latex}{\out{\hypertarget{method-get_p}{}}}
\subsection{Method \code{get_p()}}{
Function to return the conditional probability of the edge which links to
the specified child node.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChanceNode$get_p(childNode)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{childNode}}{child node to which to find probability of linking edge}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Numerical value of probability.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_modvars"></a>}}
\if{latex}{\out{\hypertarget{method-get_modvars}{}}}
\subsection{Method \code{get_modvars()}}{
Return the list of model variables associated with the node. The
model variables may be associated with costs or probabilities.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChanceNode$get_modvars()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of model variables.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ChanceNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
