% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMarkovModel.R
\docType{class}
\name{CohortMarkovModel}
\alias{CohortMarkovModel}
\title{CohortMarkovModel}
\description{
An R6 class for a Markov model with cohort simulation.
}
\details{
A class to represent a complete Markov model.
}
\author{
Andrew J. Sims \email{andrew.sims@newcastle.ac.uk}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CohortMarkovModel$new()}}
\item \href{#method-set_populations}{\code{CohortMarkovModel$set_populations()}}
\item \href{#method-cycle}{\code{CohortMarkovModel$cycle()}}
\item \href{#method-cycles}{\code{CohortMarkovModel$cycles()}}
\item \href{#method-get_statenames}{\code{CohortMarkovModel$get_statenames()}}
\item \href{#method-setDiscount}{\code{CohortMarkovModel$setDiscount()}}
\item \href{#method-tabulate_states}{\code{CohortMarkovModel$tabulate_states()}}
\item \href{#method-clone}{\code{CohortMarkovModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a Markov model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMarkovModel$new(states, Ip, discount = 0, nCyclesPerYear = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{states}}{a list of objects of type MarkovState}

\item{\code{Ip}}{Matrix of \emph{annual} rates of transition between states,
represented by a TransitionMatrix object.}

\item{\code{discount}}{The annual discount rate (percentage).}

\item{\code{nCyclesPerYear}}{The number of cycles per year.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A MarkovModel object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_populations"></a>}}
\if{latex}{\out{\hypertarget{method-set_populations}{}}}
\subsection{Method \code{set_populations()}}{
Sets the occupancy of each state. 
Calling this resets the cycle count for the model to zero.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMarkovModel$set_populations(populations)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{populations}}{A named vector of populations for
the start of the state. The names should be the state names. 
Due to R's implementation of matrix algebra, \code{populations} 
must be a numeric type and is not restricted to being an integer.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated MarkovModel object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cycle"></a>}}
\if{latex}{\out{\hypertarget{method-cycle}{}}}
\subsection{Method \code{cycle()}}{
Applies one cycle of the model, starting at zero.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMarkovModel$cycle()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Calculated values, per state.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cycles"></a>}}
\if{latex}{\out{\hypertarget{method-cycles}{}}}
\subsection{Method \code{cycles()}}{
Applies \code{nCycles} cycles of the model. The starting
populations are redistributed through the
transition probabilities and the state occupancy costs are
calculated, using function \code{cycle}. The end populations are
then fed back into the model for a further cycle and the
process is repeated. For each cycle, the state populations and
the aggregated occupancy costs are saved in one row of the
returned dataframe, with the cycle number. If the cycle count
for the model is zero when called, the first cycle evaluated
will be cycle zero, i.e. the distribution of patients to starting
states, which will include calculation of state entry costs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMarkovModel$cycles(nCycles = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nCycles}}{Number of cycles to run; default is 2.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame with cycle results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_statenames"></a>}}
\if{latex}{\out{\hypertarget{method-get_statenames}{}}}
\subsection{Method \code{get_statenames()}}{
Returns a character list of state names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMarkovModel$get_statenames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of the names of each state.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setDiscount"></a>}}
\if{latex}{\out{\hypertarget{method-setDiscount}{}}}
\subsection{Method \code{setDiscount()}}{
Set annual discount rate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMarkovModel$setDiscount(r)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r}}{Annual discount rate, expressed as a percentage.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated MarkovModel object.
Sets the annual state transition rates.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tabulate_states"></a>}}
\if{latex}{\out{\hypertarget{method-tabulate_states}{}}}
\subsection{Method \code{tabulate_states()}}{
Creates a state summary data frame suitable for printing with kable etc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMarkovModel$tabulate_states()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Ip}}{A numeric square matrix of dimension equal to the number of
states, with row and column names equal to Markov state names. Transition 
rates are from row state to column state. Exactly one element of each
row should be NA; this element is computed from the others to ensure that
the sum of 'from' probabilities is unity/}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated MarkovModel object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMarkovModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
