% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionNode.R
\docType{class}
\name{DecisionNode}
\alias{DecisionNode}
\title{DecisionNode}
\description{
An R6 class for a decision node in a decision tree
}
\details{
A class to represent a decision node in a decision tree. The node
is associated with one or more branches to child nodes.
}
\author{
Andrew J. Sims \email{andrew.sims@newcastle.ac.uk}
}
\section{Super class}{
\code{\link[rdecision:Node]{rdecision::Node}} -> \code{DecisionNode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DecisionNode$new()}}
\item \href{#method-get_choices}{\code{DecisionNode$get_choices()}}
\item \href{#method-clone}{\code{DecisionNode$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="child_nodes">}\href{../../rdecision/html/Node.html#method-child_nodes}{\code{rdecision::Node$child_nodes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="descendantNodes">}\href{../../rdecision/html/Node.html#method-descendantNodes}{\code{rdecision::Node$descendantNodes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="get_edge_label">}\href{../../rdecision/html/Node.html#method-get_edge_label}{\code{rdecision::Node$get_edge_label()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="get_label">}\href{../../rdecision/html/Node.html#method-get_label}{\code{rdecision::Node$get_label()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="get_modvars">}\href{../../rdecision/html/Node.html#method-get_modvars}{\code{rdecision::Node$get_modvars()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="getPathways">}\href{../../rdecision/html/Node.html#method-getPathways}{\code{rdecision::Node$getPathways()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="has_child_nodes">}\href{../../rdecision/html/Node.html#method-has_child_nodes}{\code{rdecision::Node$has_child_nodes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="is_same_node">}\href{../../rdecision/html/Node.html#method-is_same_node}{\code{rdecision::Node$is_same_node()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="node_type">}\href{../../rdecision/html/Node.html#method-node_type}{\code{rdecision::Node$node_type()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="sample_modvars">}\href{../../rdecision/html/Node.html#method-sample_modvars}{\code{rdecision::Node$sample_modvars()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="tabulate_modvars">}\href{../../rdecision/html/Node.html#method-tabulate_modvars}{\code{rdecision::Node$tabulate_modvars()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new decision node.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DecisionNode$new(children, choices)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{children}}{A list of Nodes which are the children of this
decision node.}

\item{\code{choices}}{A list of character stings containing the labels
of each choice associated with the decision.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new DecisionNode object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_choices"></a>}}
\if{latex}{\out{\hypertarget{method-get_choices}{}}}
\subsection{Method \code{get_choices()}}{
Return names of the choices associated with this node.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DecisionNode$get_choices()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of character strings.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DecisionNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
