% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edge.R
\docType{class}
\name{Edge}
\alias{Edge}
\title{Edge}
\description{
An R6 class to represent an edge in a decision tree
}
\details{
Edges are the formal term for paths linking nodes in a
hierarchical tree. It is not intended that package users creating models
should instantiate the `Edge` class. Instead, it is included in the
package as a convenience class used in the construction and traversal
of decision trees by the package methods themselves. In this case
objects of type `Edge` are used to collect together information relating to 
the same edge, i.e. a label, a cost and a conditional probability.
}
\author{
Andrew J. Sims \email{andrew.sims5@nhs.net}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Edge$new()}}
\item \href{#method-getToNode}{\code{Edge$getToNode()}}
\item \href{#method-getLabel}{\code{Edge$getLabel()}}
\item \href{#method-getCost}{\code{Edge$getCost()}}
\item \href{#method-setCost}{\code{Edge$setCost()}}
\item \href{#method-getP}{\code{Edge$getP()}}
\item \href{#method-setP}{\code{Edge$setP()}}
\item \href{#method-clone}{\code{Edge$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an object of type `Edge`.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Edge$new(fromNode, toNode, label, cost = 0, p = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fromNode}}{Node nearest the root to which the edge connects.}

\item{\code{toNode}}{Node nearest the leaf to which the edge connects.}

\item{\code{label}}{Character string containing the edge label.}

\item{\code{cost}}{Cost associated with traversing the edge; ModelVariable.}

\item{\code{p}}{Probability of traversing the edge conditional on
having reached `fromNode`; ModelVariable.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Edge` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getToNode"></a>}}
\if{latex}{\out{\hypertarget{method-getToNode}{}}}
\subsection{Method \code{getToNode()}}{
Access toNode.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Edge$getToNode()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
`Node` to which the edge leads.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLabel"></a>}}
\if{latex}{\out{\hypertarget{method-getLabel}{}}}
\subsection{Method \code{getLabel()}}{
Access label.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Edge$getLabel()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Label of the edge; character string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCost"></a>}}
\if{latex}{\out{\hypertarget{method-getCost}{}}}
\subsection{Method \code{getCost()}}{
Access cost.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Edge$getCost()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Cost associated with traversing the edge; numeric.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setCost"></a>}}
\if{latex}{\out{\hypertarget{method-setCost}{}}}
\subsection{Method \code{setCost()}}{
Set the cost associated with traversing the edge.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Edge$setCost(c)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{c}}{Cost as a numeric value.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getP"></a>}}
\if{latex}{\out{\hypertarget{method-getP}{}}}
\subsection{Method \code{getP()}}{
Get the conditional probability of traversing the edge.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Edge$getP()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Conditional probability; numeric.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setP"></a>}}
\if{latex}{\out{\hypertarget{method-setP}{}}}
\subsection{Method \code{setP()}}{
Set conditional probability of traversing the edge
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Edge$setP(p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{Conditional probability value; a numeric in the range [0,1].}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated `edge` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Edge$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
