% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeafNode.R
\docType{class}
\name{LeafNode}
\alias{LeafNode}
\title{LeafNode}
\description{
An R6 class for a leaf node in a decision tree
}
\details{
A LefNode is synomymous with a clinical outcome, or end
point of a decision tree. Paths through the tree must end on leaf
nodes.
}
\author{
Andrew J. Sims \email{andrew.sims5@nhs.net}
}
\section{Super class}{
\code{\link[rdecision:Node]{rdecision::Node}} -> \code{LeafNode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LeafNode$new()}}
\item \href{#method-getPathway}{\code{LeafNode$getPathway()}}
\item \href{#method-getUtility}{\code{LeafNode$getUtility()}}
\item \href{#method-clone}{\code{LeafNode$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="childNodes">}\href{../../rdecision/html/Node.html#method-childNodes}{\code{rdecision::Node$childNodes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="descendantNodes">}\href{../../rdecision/html/Node.html#method-descendantNodes}{\code{rdecision::Node$descendantNodes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="getCost">}\href{../../rdecision/html/Node.html#method-getCost}{\code{rdecision::Node$getCost()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="getLabel">}\href{../../rdecision/html/Node.html#method-getLabel}{\code{rdecision::Node$getLabel()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="getModelVariables">}\href{../../rdecision/html/Node.html#method-getModelVariables}{\code{rdecision::Node$getModelVariables()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="getP">}\href{../../rdecision/html/Node.html#method-getP}{\code{rdecision::Node$getP()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="hasChildNodes">}\href{../../rdecision/html/Node.html#method-hasChildNodes}{\code{rdecision::Node$hasChildNodes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="isSameNode">}\href{../../rdecision/html/Node.html#method-isSameNode}{\code{rdecision::Node$isSameNode()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="nodeType">}\href{../../rdecision/html/Node.html#method-nodeType}{\code{rdecision::Node$nodeType()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="sample">}\href{../../rdecision/html/Node.html#method-sample}{\code{rdecision::Node$sample()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="tabulateModelVariables">}\href{../../rdecision/html/Node.html#method-tabulateModelVariables}{\code{rdecision::Node$tabulateModelVariables()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new `LeafNode` object; synonymous with a clinical outcome.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeafNode$new(name, utility = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character string; a label for the leaf node which is
synonymous with the name of the root-to-leaf pathway}

\item{\code{utility}}{the preference or value that a user associates with
a given health state (range 0 to 1).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `LeafNode` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getPathway"></a>}}
\if{latex}{\out{\hypertarget{method-getPathway}{}}}
\subsection{Method \code{getPathway()}}{
Return the label of the leaf node; the name of the clinical outcome.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeafNode$getPathway()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Name of the clinical outcome; character string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getUtility"></a>}}
\if{latex}{\out{\hypertarget{method-getUtility}{}}}
\subsection{Method \code{getUtility()}}{
Return the utility associated with the clinical outcome.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeafNode$getUtility()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Utility (numeric value)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeafNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
