% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeafNode.R
\docType{class}
\name{LeafNode}
\alias{LeafNode}
\title{LeafNode}
\description{
An R6 class for a leaf node in a decision tree representing a clinical state.
}
\details{
It represents a state of being, and is associated with an annual cost,
an incremental utility and a benefit. It inherits from class Node so that it 
can be part of a decision tree as a leaf node with no children.
}
\author{
Andrew J. Sims \email{andrew.sims@newcastle.ac.uk}
}
\section{Super class}{
\code{\link[rdecision:Node]{rdecision::Node}} -> \code{LeafNode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LeafNode$new()}}
\item \href{#method-cost}{\code{LeafNode$cost()}}
\item \href{#method-utility}{\code{LeafNode$utility()}}
\item \href{#method-benefit}{\code{LeafNode$benefit()}}
\item \href{#method-clone}{\code{LeafNode$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="label">}\href{../../rdecision/html/Node.html#method-label}{\code{rdecision::Node$label()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="rdecision" data-topic="Node" data-id="type">}\href{../../rdecision/html/Node.html#method-type}{\code{rdecision::Node$type()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{LeafNode} object; synonymous with a clinical outcome.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeafNode$new(
  name,
  cost = 0,
  utility = 1,
  benefit = 0,
  interval = as.difftime(365.25, units = "days")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character string; a label for the state.}

\item{\code{cost}}{The cost of being in the state over the interval.}

\item{\code{utility}}{The incremental utility that a user associates with
being in the health state (range -Inf to 1) for the interval. Intended
for use with cost benefit analysis.}

\item{\code{benefit}}{The financial benefit accruing from reaching the node.
Intended for use with cost-benefit analysis. Normally only one of
utility or benefit are set.}

\item{\code{interval}}{The time interval, over which the \code{cost} and
\code{utility} parameters apply, expressed as an R \code{difftime} object;
default 1 year.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{LeafNode} object
#' @description 
#' Return the label of the state; the name of the clinical outcome.
#' @return Name of the clinical outcome or state; character string.
get_name = function() {
  return(private$name)
},
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cost"></a>}}
\if{latex}{\out{\hypertarget{method-cost}{}}}
\subsection{Method \code{cost()}}{
Return the cost incurred by being in the state for the interval.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeafNode$cost()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Cost, as a numeric value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-utility"></a>}}
\if{latex}{\out{\hypertarget{method-utility}{}}}
\subsection{Method \code{utility()}}{
Return the incremental utility associated with being in the state for
the interval.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeafNode$utility()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Incremental utility (numeric value).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-benefit"></a>}}
\if{latex}{\out{\hypertarget{method-benefit}{}}}
\subsection{Method \code{benefit()}}{
Return the benefit (payoff) associated with being in the state for
the interval.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeafNode$benefit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Benefit (numeric value).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeafNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
