% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovState.R
\docType{class}
\name{MarkovState}
\alias{MarkovState}
\title{MarkovState}
\description{
An R6 class for a state in a Markov model
}
\details{
A class to represent a single state in a Markov model.
}
\author{
Andrew J. Sims \email{andrew.sims@newcastle.ac.uk}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MarkovState$new()}}
\item \href{#method-get_name}{\code{MarkovState$get_name()}}
\item \href{#method-hasCycleLimit}{\code{MarkovState$hasCycleLimit()}}
\item \href{#method-getCycleLimit}{\code{MarkovState$getCycleLimit()}}
\item \href{#method-setAnnualCost}{\code{MarkovState$setAnnualCost()}}
\item \href{#method-getAnnualCost}{\code{MarkovState$getAnnualCost()}}
\item \href{#method-setEntryCost}{\code{MarkovState$setEntryCost()}}
\item \href{#method-getEntryCost}{\code{MarkovState$getEntryCost()}}
\item \href{#method-clone}{\code{MarkovState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Markov state object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$new(
  name,
  annualCost = 0,
  entryCost = 0,
  hasCycleLimit = F,
  cycleLimit = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the state (character string).}

\item{\code{annualCost}}{The annual cost of state occupancy.}

\item{\code{entryCost}}{The cost to enter the state.}

\item{\code{hasCycleLimit}}{Whether the state has a cycle limit; logical.}

\item{\code{cycleLimit}}{The maximum length of stay (1 for cohort tunnel states); numeric.}

\item{\code{isTemporaryState}}{Logical; TRUE if the state is a tunnel state.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of type MarkovState.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-get_name}{}}}
\subsection{Method \code{get_name()}}{
Accessor function to retrieve the state name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$get_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
State name.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hasCycleLimit"></a>}}
\if{latex}{\out{\hypertarget{method-hasCycleLimit}{}}}
\subsection{Method \code{hasCycleLimit()}}{
Accessor function to identify if the state has a cycle limit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$hasCycleLimit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE if temporary state; false otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCycleLimit"></a>}}
\if{latex}{\out{\hypertarget{method-getCycleLimit}{}}}
\subsection{Method \code{getCycleLimit()}}{
Accessor function to find the cycle limit.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$getCycleLimit()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The cycle limit; integer.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setAnnualCost"></a>}}
\if{latex}{\out{\hypertarget{method-setAnnualCost}{}}}
\subsection{Method \code{setAnnualCost()}}{
Sets the annual cost of state occupancy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$setAnnualCost(annualCost)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{annualCost}}{Annual cost of occupying the state; numeric.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated MarkovState object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAnnualCost"></a>}}
\if{latex}{\out{\hypertarget{method-getAnnualCost}{}}}
\subsection{Method \code{getAnnualCost()}}{
Gets the annual cost of state occupancy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$getAnnualCost()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Annual cost; numeric.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setEntryCost"></a>}}
\if{latex}{\out{\hypertarget{method-setEntryCost}{}}}
\subsection{Method \code{setEntryCost()}}{
Sets the entry cost of state occupancy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$setEntryCost(entryCost)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entryCost}}{Cost to enter the state; numeric.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Updated MarkovState object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getEntryCost"></a>}}
\if{latex}{\out{\hypertarget{method-getEntryCost}{}}}
\subsection{Method \code{getEntryCost()}}{
Gets the entry cost of state occupancy.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$getEntryCost()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Entry cost; numeric.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
