% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdecision-package.R
\docType{package}
\name{rdecision-package}
\alias{rdecision}
\alias{rdecision-package}
\title{rdecision: Decision Analytic Modelling in Health Economics.}
\description{
The goal of `rdecision` is to provide methods for assessing health care
interventions using cohort models (decision trees and semi-Markov models)
which can be constructed using only a few lines of R code. Mechanisms
are provided for associating an uncertainty distribution with each source
variable and for ensuring transparency of the mathematical relationships
between
variables. The package terminology follows Briggs \emph{et al} "Decision
Modelling for Health Economic Evaluation" (2006, ISBN:978-0-19-852662-9).
}
\author{
\strong{Maintainer}: Andrew Sims \email{andrew.sims@newcastle.ac.uk} (\href{https://orcid.org/0000-0002-9553-7278}{ORCID})

Authors:
\itemize{
  \item Kim Fairbairn \email{kim.fairbairn@nhs.net} (\href{https://orcid.org/0000-0001-5108-6279}{ORCID})
  \item Paola Cognigni \email{paola.cognigni@nhs.net} (\href{https://orcid.org/0000-0001-5418-3103}{ORCID})
}

}
\keyword{internal}
