% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdecision.R
\docType{package}
\name{rdecision}
\alias{rdecision}
\title{rdecision: Decision Analytic Modelling in Health Economics.}
\description{
Classes and functions for modelling health care interventions 
using cohort models (decision trees). Mechanisms are provided for 
associating an uncertainty distribution with each source variable and for
ensuring transparency of the mathematical relationships between variables.
The package terminology follows Briggs et al "Decision Modelling for Health
Economic Evaluation" (2006, ISBN:978-0-19-852662-9).
}
