% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdecision.R
\docType{package}
\name{rdecision}
\alias{rdecision}
\title{rdecision: Decision Analytic Modelling in Health Economics.}
\description{
Classes and functions for modelling healthcare interventions using
cohort models (decision trees, Markov models and extended Markov models).
It draws on terminology from Briggs, Claxton and Sculpher, "Decision
Modelling for Health Economic Evaluation", Oxford University Press, 2006.
}
